requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
djangorestframework==3.16.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
